+++ Main/Main/cat-service/src/main/java/Entities/Color.java
@@ -1,0 +1,6 @@
+package Entities;
+
+public enum Color {
+    BLACK,
+    WHITE
+}
