+++ Main/Main/cat-service/src/main/java/Repository/CatRepository.java
@@ -1,12 +1,10 @@
 package Repository;
 
 
-import Entities.User;
+import Entities.Cat;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends CrudRepository<User, Long> {
-    User findByName(String name);
-
+public interface CatRepository extends CrudRepository<Cat, Long> {
 }
