+++ Main/Main/cat-service/src/main/java/Service/CatService.java
@@ -1,0 +1,4 @@
+package Service;
+
+public class CatService {
+}
