--- Main/Main/cat-service/src/main/java/kkkombinator/Entities/Color.java
@@ -1,6 +1,0 @@
-package kkkombinator.Entities;
-
-public enum Color {
-    BLACK,
-    WHITE
-}
