--- Main/Main/cat-service/src/main/java/kkkombinator/Repository/CatRepository.java
@@ -1,10 +1,0 @@
-package kkkombinator.Repository;
-
-
-import kkkombinator.Entities.Cat;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CatRepository extends CrudRepository<Cat, Long> {
-}
