--- Main/Main/cat-service/src/main/java/kkkombinator/Service/ColorMapper.java
@@ -1,20 +1,0 @@
-package kkkombinator.Service;
-
-
-import kkkombinator.Entities.Color;
-import org.springframework.stereotype.Service;
-
-@Service
-public class ColorMapper {
-    public Color getColorFromString(String color) throws IllegalArgumentException{
-        return Color.valueOf(color);
-    }
-
-    public String getStringFromColor(Color color) {
-
-        return switch (color) {
-            case Color.BLACK -> "BLACK";
-            case Color.WHITE -> "WHITE";
-        };
-    }
-}
