--- Main/Main/common-dto/src/main/java/kkkombinator/Color.java
@@ -1,6 +1,0 @@
-package kkkombinator;
-
-public enum Color {
-    BLACK,
-    WHITE
-}
