--- Main/Main/common-dto/src/main/java/kkkombinator/KafkaTopics.java
@@ -1,7 +1,0 @@
-package kkkombinator;
-
-public class KafkaTopics {
-    public static final String CATS_TOPIC = "cats-topic";
-    public static final String USERS_TOPIC = "users-topic";
-    public static final String CATS_REQUEST_TOPIC = "cats-request-topic";
-}
