--- Main/Main/common-dto/src/main/java/kkkombinator/Requests/UserRequest.java
@@ -1,13 +1,0 @@
-package kkkombinator.Requests;
-
-import kkkombinator.UserDTO;
-
-public class UserRequest {
-    private String action;
-    private UserDTO dto;
-
-    public UserRequest(String action, UserDTO dto) {
-        this.action = action;
-        this.dto = dto;
-    }
-}
