--- Main/Main/common-dto/src/main/resources/application.properties
@@ -1,16 +1,0 @@
-spring.application.name=lab3
-server.port=8080
-
-spring.thymeleaf.cache=false
-spring.thymeleaf.enabled=true
-spring.thymeleaf.prefix=classpath:/templates/
-spring.thymeleaf.suffix=.html
-
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.show-sql=true
-spring.datasource.driver-class-name=org.postgresql.Driver
-spring.datasource.url=jdbc:postgresql://localhost:5431/postgres
-spring.datasource.username=postgres
-spring.datasource.password=postgres
-
-logging.level.org.springframework.security=TRACE
