+++ Main/Main/user-service/src/main/java/Entities/Role.java
@@ -1,0 +1,34 @@
+package Entities;
+
+import jakarta.persistence.*;
+import lombok.Data;
+import lombok.RequiredArgsConstructor;
+
+import java.util.Collection;
+
+@Entity
+@Data
+@RequiredArgsConstructor
+public class Role {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    public Role(String name) {
+        this.name = name;
+    }
+
+    private String name;
+    @ManyToMany(mappedBy = "roles")
+    private Collection<User> users;
+
+    @ManyToMany
+    @JoinTable(
+            name = "rolesPrivileges",
+            joinColumns = @JoinColumn(
+                    name = "roleId", referencedColumnName = "id"),
+            inverseJoinColumns = @JoinColumn(
+                    name = "privilegeId", referencedColumnName = "id"))
+    private Collection<Privilege> privileges;
+}
