+++ Main/Main/user-service/src/main/java/Repository/PrivilegeRepository.java
@@ -1,7 +1,7 @@
 package Repository;
 
 
-import kkkombinator.Entities.Privilege;
+import Entities.Privilege;
 import org.springframework.data.repository.CrudRepository;
 
 public interface PrivilegeRepository extends CrudRepository<Privilege, Long> {
