+++ Main/Main/user-service/src/main/java/Repository/RoleRepository.java
@@ -1,7 +1,7 @@
 package Repository;
 
 
-import kkkombinator.Entities.Role;
+import Entities.Role;
 import org.springframework.data.repository.CrudRepository;
 
 public interface RoleRepository extends CrudRepository<Role, Long> {
