+++ Main/Main/user-service/src/main/java/Service/SetupDataLoader.java
@@ -3,19 +3,16 @@ package Service;
 import Repository.PrivilegeRepository;
 import Repository.RoleRepository;
 import Repository.UserRepository;
-import kkkombinator.Entities.Privilege;
-import kkkombinator.Entities.Role;
-import kkkombinator.Entities.User;
+import Entities.Privilege;
+import Entities.Role;
+import Entities.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationListener;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.event.ContextRefreshedEvent;
 import org.springframework.security.access.hierarchicalroles.RoleHierarchy;
 import org.springframework.security.access.hierarchicalroles.RoleHierarchyImpl;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.access.expression.DefaultWebSecurityExpressionHandler;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
