+++ Main/Main/user-service/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.application.name=lab3
+spring.application.name=user-service
 server.port=8080
 
 spring.thymeleaf.cache=false
@@ -14,3 +14,12 @@ spring.datasource.username=postgres
 spring.datasource.password=postgres
 
 logging.level.org.springframework.security=TRACE
+
+spring.kafka.bootstrap-servers = localhost:/9092
+
+spring.kafka.userCreateTopic=users-create
+spring.kafka.userFindAllTopic=users-find-all
+spring.kafka.userUpdateTopic=users-update
+spring.kafka.userDeleteTopic=users-delete
+
+spring.kafka.consumer.group-id = 1
