+++ Main/Main/web-gateway/pom.xml
@@ -20,11 +20,6 @@
 
     <dependencies>
         <dependency>
-            <groupId>kkkombinator</groupId>
-            <artifactId>common-dto</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-web</artifactId>
         </dependency>
