+++ Main/Main/web-gateway/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.application.name=lab3
+spring.application.name=web-gateway
 server.port=8080
 
 spring.thymeleaf.cache=false
@@ -14,3 +14,22 @@ spring.datasource.username=postgres
 spring.datasource.password=postgres
 
 logging.level.org.springframework.security=TRACE
+
+spring.kafka.bootstrap-servers = localhost:/9092
+
+spring.kafka.consumer.group-id = 1
+
+
+spring.kafka.catCreateTopic=cats-create
+spring.kafka.catFindAllTopic=cats-find-all
+spring.kafka.catUpdateTopic=cats-update
+spring.kafka.catDeleteTopic=cats-delete
+
+spring.kafka.userCreateTopic=users-create
+spring.kafka.userFindAllTopic=users-find-all
+spring.kafka.userUpdateTopic=users-update
+spring.kafka.userDeleteTopic=users-delete
+
+spring.kafka.globalReplicationFactor = 1
+spring.kafka.globalPartition = 1
+sptring.kafka.timeout=5
